Refactor shape vertex/index management & simplify updates.
