+++ sunfish-core/src/ui/shapes.rs
@@ -1,4 +1,3 @@
-use std::collections::HashSet;
 use std::mem;
 
 use bytemuck::{Pod, Zeroable};
@@ -17,8 +16,8 @@ use lyon::tessellation::{
 use lyon::tessellation::{FillVertex, StrokeVertex};
 use serde::Deserialize;
 
-use crate::ui::buffers;
 use crate::ui::coords::Rect;
+use crate::ui::packed_shapes;
 
 pub type Buffers = tessellation::VertexBuffers<ShapeVertex, u16>;
 
@@ -165,7 +164,13 @@ pub struct ShapeVertex {
 }
 
 impl ShapeVertex {
-    pub fn descriptor<'a>() -> wgpu::VertexBufferDescriptor<'a> {
+    pub fn new(position: [f32; 3], color: [f32; 3]) -> Self {
+        Self { position, color }
+    }
+}
+
+impl packed_shapes::Vertex for ShapeVertex {
+    fn descriptor<'a>() -> wgpu::VertexBufferDescriptor<'a> {
         wgpu::VertexBufferDescriptor {
             stride: mem::size_of::<Self>() as wgpu::BufferAddress,
             step_mode: wgpu::InputStepMode::Vertex,
@@ -182,9 +187,6 @@ impl ShapeVertex {
                 },
             ],
         }
-    }
-    pub fn new(position: [f32; 3], color: [f32; 3]) -> Self {
-        Self { position, color }
     }
 }
 
@@ -215,268 +217,6 @@ impl<'a> lyon::tessellation::FillVertexConstructor<ShapeVertex> for ShapeVertexB
             color: self.color,
         }
     }
-}
-
-/// A Shape captures the CPU side of a single polygon with a
-/// max number of vertices known ahead of time (to simplify
-/// GPU-side memory management).
-pub struct Shape {
-    vertices: Vec<ShapeVertex>,
-    indices: Vec<u16>,
-    /// Max vertices and indices allocated for this shape.
-    max_v_count: usize,
-    max_i_count: usize,
-}
-
-impl Shape {
-    pub fn new(
-        vertices: Vec<ShapeVertex>,
-        indices: Vec<u16>,
-        max_v_count: usize,
-        max_i_count: usize,
-    ) -> Self {
-        Self {
-            vertices,
-            indices,
-            max_v_count,
-            max_i_count,
-        }
-    }
-
-    pub fn from_lyon(shapes: Buffers, max_v_count: usize, max_i_count: usize) -> Self {
-        Shape {
-            vertices: shapes.vertices,
-            indices: shapes.indices,
-            max_v_count,
-            max_i_count,
-        }
-    }
-
-    fn update(&mut self, vertices: &[ShapeVertex], indices: &[u16]) {
-        // TODO: Copy the slices in place.
-        let vertices = if vertices.len() > self.max_v_count {
-            // TODO: Should we just truncate and log?
-            log::warn!("Shape::update received vertex buffer larger than max_v_count");
-            &vertices[..self.max_v_count]
-        } else {
-            vertices
-        };
-
-        let indices = if indices.len() > self.max_i_count {
-            // TODO: Should we just truncate and log?
-            log::warn!("Shape::update received index buffer larger than max_i_count");
-            &indices[..self.max_i_count]
-        } else {
-            indices
-        };
-
-        // Replace them in-memory.
-        self.vertices.resize(vertices.len(), ShapeVertex::zeroed());
-        self.vertices.clone_from_slice(vertices);
-
-        // TODO: Maybe Shapes and BoundShapes should be merged, then we can
-        // look directly at the offset from ind_ranges.
-        self.indices.resize(indices.len(), 0);
-        self.indices.clone_from_slice(indices);
-    }
-}
-
-pub struct Shapes {
-    shapes: Vec<Shape>,
-    shapes_to_update: HashSet<usize>,
-}
-
-impl Shapes {
-    pub fn with_capacity(shape_count: usize) -> Self {
-        Self {
-            shapes: Vec::with_capacity(shape_count),
-            shapes_to_update: HashSet::with_capacity(shape_count),
-        }
-    }
-
-    pub fn add(&mut self, shape: Shape) -> usize {
-        let index = self.shapes.len();
-        self.shapes.push(shape);
-        index
-    }
-
-    pub fn update(&mut self, index: usize, vertices: &[ShapeVertex], indices: &[u16]) {
-        self.shapes_to_update.insert(index);
-        if let Some(shape) = self.shapes.get_mut(index) {
-            shape.update(vertices, indices)
-        }
-    }
-}
-
-struct VerRanges(Vec<std::ops::Range<u32>>);
-struct IndRanges(Vec<std::ops::Range<u32>>);
-
-pub struct BoundShapes {
-    pub shape_render_pipeline: wgpu::RenderPipeline,
-    buffers: buffers::VertexBuffers<ShapeVertex>,
-    ver_ranges: VerRanges,
-    ind_ranges: IndRanges,
-    shadow_vertices: Vec<ShapeVertex>,
-    shadow_indices: Vec<u16>,
-}
-
-impl BoundShapes {
-    pub fn new(
-        device: &wgpu::Device,
-        swapchain_format: &wgpu::TextureFormat,
-        shapes: &Shapes,
-    ) -> Self {
-        let (ver_ranges, ind_ranges, ver_buf, ind_buf) = Self::pack_shapes(shapes);
-        let buffers = buffers::VertexBuffers::new(device, &ver_buf, &ind_buf);
-
-        let shape_vs_module =
-            device.create_shader_module(wgpu::include_spirv!("shader_shape.vert.spv"));
-        let shape_fs_module =
-            device.create_shader_module(wgpu::include_spirv!("shader_shape.frag.spv"));
-
-        let shape_render_pipeline_layout =
-            device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-                label: Some("shape_render_pipeline_layout"),
-                bind_group_layouts: &[],
-                push_constant_ranges: &[],
-            });
-
-        let shape_render_pipeline =
-            device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-                label: Some("shape_render_pipeline"),
-                layout: Some(&shape_render_pipeline_layout),
-                vertex_stage: wgpu::ProgrammableStageDescriptor {
-                    module: &shape_vs_module,
-                    entry_point: "main",
-                },
-                fragment_stage: Some(wgpu::ProgrammableStageDescriptor {
-                    module: &shape_fs_module,
-                    entry_point: "main",
-                }),
-                // Use the default rasterizer state: no culling, no depth bias
-                rasterization_state: None,
-                primitive_topology: wgpu::PrimitiveTopology::TriangleList,
-                color_states: &[(*swapchain_format).into()],
-                depth_stencil_state: None,
-                vertex_state: wgpu::VertexStateDescriptor {
-                    index_format: wgpu::IndexFormat::Uint16,
-                    vertex_buffers: &[ShapeVertex::descriptor()],
-                },
-                sample_count: 1,
-                sample_mask: !0,
-                alpha_to_coverage_enabled: false,
-            });
-        BoundShapes {
-            shape_render_pipeline,
-            buffers,
-            ver_ranges,
-            ind_ranges,
-            shadow_vertices: ver_buf,
-            shadow_indices: ind_buf,
-        }
-    }
-
-    fn pack_shapes(shapes: &Shapes) -> (VerRanges, IndRanges, Vec<ShapeVertex>, Vec<u16>) {
-        // Compute total buffer size.
-        let tot_ver_buf: usize = shapes.shapes.iter().map(|shape| shape.max_v_count).sum();
-        let tot_ind_buf: usize = shapes.shapes.iter().map(|shape| shape.max_i_count).sum();
-
-        // Round up to nearest 4. TODO: figure out how to get the 4.0
-        let tot_ind_buf = (((tot_ind_buf as f32 / 4.0).ceil()) * 4.0) as usize;
-
-        let mut ver_buf = vec![ShapeVertex::zeroed(); tot_ver_buf];
-        let mut ind_buf = vec![0u16; tot_ind_buf];
-
-        let mut ver_offset = 0;
-        let mut ver_last_offset = 0;
-
-        let mut ind_offset = 0;
-        let mut ind_last_offset = 0;
-
-        let mut ver_ranges = Vec::with_capacity(shapes.shapes.len());
-        let mut ind_ranges = Vec::with_capacity(shapes.shapes.len());
-
-        for shape in &shapes.shapes {
-            let ver_size = shape.max_v_count;
-            let ind_size = shape.max_i_count;
-
-            // Copy vertex and index data; note that the sizes of
-            // them will be <= max_{i,v}_count
-            //
-            // copy into the buffer
-            ver_buf[ver_last_offset..ver_last_offset + shape.vertices.len()]
-                .clone_from_slice(&shape.vertices);
-
-            // We have to offset the indices to account for previous vertices.
-            for (i, vertex_index) in shape.indices.iter().enumerate() {
-                ind_buf[i + ind_last_offset] = vertex_index + ver_last_offset as u16;
-            }
-
-            ver_offset += ver_size;
-            ind_offset += ind_size;
-
-            ver_ranges.push(ver_last_offset as u32..ver_offset as u32);
-            ind_ranges.push(ind_last_offset as u32..ind_offset as u32);
-
-            ver_last_offset = ver_offset;
-            ind_last_offset = ind_offset;
-        }
-        (
-            VerRanges(ver_ranges),
-            IndRanges(ind_ranges),
-            ver_buf,
-            ind_buf,
-        )
-    }
-
-    fn update_shadow_buffer(&mut self, shapes: &mut Shapes) {
-        // TODO:
-        // - update the ind_ranges
-        for shape_index in shapes.shapes_to_update.drain() {
-            let shape = &shapes.shapes[shape_index];
-            let ver_offset = self.ver_ranges.0[shape_index].start;
-            let ver_offset_usize = ver_offset as usize;
-            let ver_size = shape.vertices.len();
-
-            self.shadow_vertices[ver_offset_usize..ver_offset_usize + ver_size]
-                .copy_from_slice(&shape.vertices);
-
-            let start = self.ver_ranges.0[shape_index].start;
-            self.ver_ranges.0[shape_index].end = start + ver_size as u32;
-
-            // =======
-            // Indices:
-            // =======
-            let ind_offset = self.ind_ranges.0[shape_index].start as usize;
-            let ind_size = shape.indices.len();
-
-            // We have to offset the indices to account for previous vertices.
-            for (i, vertex_index) in shape.indices.iter().enumerate() {
-                self.shadow_indices[i + ind_offset] = vertex_index + ver_offset as u16;
-            }
-
-            let start = self.ind_ranges.0[shape_index].start;
-            self.ind_ranges.0[shape_index].end = start + ind_size as u32;
-        }
-    }
-}
-
-pub fn render<'a>(
-    bound_shapes: &'a BoundShapes,
-    mut rpass: wgpu::RenderPass<'a>,
-) -> wgpu::RenderPass<'a> {
-    rpass.set_pipeline(&bound_shapes.shape_render_pipeline);
-
-    rpass.set_vertex_buffer(0, bound_shapes.buffers.vertices.buf.slice(..));
-    rpass.set_index_buffer(bound_shapes.buffers.indices.buf.slice(..));
-
-    for range in &bound_shapes.ind_ranges.0 {
-        if !range.is_empty() {
-            rpass.draw_indexed(range.clone(), 0, 0..1);
-        }
-    }
-
-    rpass
 }
 
 #[derive(Debug)]
@@ -561,124 +301,6 @@ impl Arc {
         }
 
         buffers
-    }
-}
-
-pub fn update(
-    device: &wgpu::Device,
-    shapes: &mut Shapes,
-    bound_shapes: &mut BoundShapes,
-    staging_belt: &mut wgpu::util::StagingBelt,
-    encoder: &mut wgpu::CommandEncoder,
-) {
-    _delta_update(device, shapes, bound_shapes, staging_belt, encoder);
-}
-
-pub fn _update_whole_copy(
-    device: &wgpu::Device,
-    shapes: &mut Shapes,
-    bound_shapes: &mut BoundShapes,
-    staging_belt: &mut wgpu::util::StagingBelt,
-    encoder: &mut wgpu::CommandEncoder,
-) {
-    // We have to update the entire buffer each time.
-    //
-    // So, to minimize copies, the bound shapes object keeps shadow vectors
-    // of the vertices and indices.
-    bound_shapes.update_shadow_buffer(shapes);
-
-    // let (ver_ranges, ind_ranges, ver_buf, ind_buf) = BoundShapes::pack_shapes(shapes);
-    // bound_shapes.ver_ranges = ver_ranges;
-    // bound_shapes.ind_ranges = ind_ranges;
-
-    // Now copy the whole chunk
-    let ver_elm_size = bound_shapes.buffers.vertices.element_size() as u64;
-    let ver_buf_size = bound_shapes.buffers.vertices.size as u64 * ver_elm_size;
-    staging_belt
-        .write_buffer(
-            encoder,
-            &bound_shapes.buffers.vertices.buf,
-            0,
-            wgpu::BufferSize::new(ver_buf_size).unwrap(), //size
-            device,
-        )
-        .copy_from_slice(bytemuck::cast_slice(&bound_shapes.shadow_vertices));
-
-    // Indices
-    let ind_elm_size = bound_shapes.buffers.indices.element_size() as u64;
-    let ind_buf_size = bound_shapes.buffers.indices.size as u64 * ind_elm_size;
-    staging_belt
-        .write_buffer(
-            encoder,
-            &bound_shapes.buffers.indices.buf,
-            0,
-            wgpu::BufferSize::new(ind_buf_size).unwrap(), //size
-            device,
-        )
-        .copy_from_slice(bytemuck::cast_slice(&bound_shapes.shadow_indices));
-}
-
-pub fn _delta_update(
-    device: &wgpu::Device,
-    shapes: &mut Shapes,
-    bound_shapes: &mut BoundShapes,
-    staging_belt: &mut wgpu::util::StagingBelt,
-    encoder: &mut wgpu::CommandEncoder,
-) {
-    // TODO:
-    // - update the ind_ranges
-    for shape_index in shapes.shapes_to_update.drain() {
-        let shape = &shapes.shapes[shape_index];
-
-        let ver_offset = bound_shapes.ver_ranges.0[shape_index].start as u64;
-        let ver_size = shape.vertices.len() as u64;
-
-        let ind_offset = bound_shapes.ind_ranges.0[shape_index].start as u64;
-        let ind_size = shape.indices.len() as u64;
-
-        // TODO: These calls dig into the guts of buffers; could probably
-        // benefit from a refactor.
-
-        // Update vertices.
-        let ver_elm_size = bound_shapes.buffers.vertices.element_size() as u64;
-        let ver_buf_size = ver_elm_size * ver_size as u64;
-        if ver_buf_size > 0 {
-            staging_belt
-                .write_buffer(
-                    encoder,
-                    &bound_shapes.buffers.vertices.buf,
-                    ver_offset * ver_elm_size,
-                    wgpu::BufferSize::new(ver_buf_size).unwrap(),
-                    device,
-                )
-                .copy_from_slice(bytemuck::cast_slice(&shape.vertices));
-        }
-        let start = bound_shapes.ver_ranges.0[shape_index].start;
-        bound_shapes.ver_ranges.0[shape_index].end = start + ver_size as u32;
-
-        // Update indices.
-        let ind_elm_size = bound_shapes.buffers.indices.element_size() as u64;
-        let ind_buf_size = ind_elm_size * ind_size as u64;
-        if ind_buf_size > 0 {
-            let mut indices_offset = vec![0u16; shape.indices.len()];
-            // We have to offset the indices to account for previous vertices.
-            for (i, vertex_index) in shape.indices.iter().enumerate() {
-                indices_offset[i] = vertex_index + ver_offset as u16;
-            }
-
-            staging_belt
-                .write_buffer(
-                    encoder,
-                    &bound_shapes.buffers.indices.buf,
-                    ind_offset * ind_elm_size,                    //offset
-                    wgpu::BufferSize::new(ind_buf_size).unwrap(), //size
-                    device,
-                )
-                .copy_from_slice(bytemuck::cast_slice(&indices_offset));
-        }
-
-        let start = bound_shapes.ind_ranges.0[shape_index].start;
-        bound_shapes.ind_ranges.0[shape_index].end = start + ind_size as u32;
     }
 }
